+++ CHANGELOG.md
@@ -6,7 +6,7 @@ N/A
 
 ### Added
 
-N/A
+- New `Swarm.Tracker.handoff/2` function which moves all processes of a worker to the remaining ones, so the worker node can be shut down gracefully [#83](https://github.com/bitwalker/swarm/pull/83).
 
 ### Removed